+++ GameCube/subrel/seed/include/rando/linkHouseSign.h
