+++ GameCube/subrel/seed/source/rando/linkHouseSign.cpp
@@ -1,4 +1,4 @@
-#include "linkHouseSign.h"
+#include "rando/linkHouseSign.h"
 #include "rando/seed.h"
 #include "tp/resource.h"
 #include "cxx.h"
